app/requirements.txt
celery==5.3.0
Django==4.2.2
django-redis==5.2.0
djangorestframework==3.14.0
numpy==1.24.3
pandas==2.0.2
redis==4.5.5
gunicorn==20.1.0
django-cors-headers==4.0.0
psycopg2-binary==2.9.6